2023/07/01 10:59

+++ jotdown/src/main/java/com/example/jotdown/MainActivity.java
@@ -238,7 +238,7 @@ public class MainActivity extends AppCompatActivity
     public void onItemClick(View view, int position) {
         Log.d(TAG, "onItemClick: id="+nodesArray.get(position)._id);
         Bundle bundle=new Bundle();                 //创建一个包裹
-        bundle.putInt("_id",nodesArray.get(position)._id);
+        bundle.putLong("_id",nodesArray.get(position)._id);
         Intent intent=new Intent(this,ModifyNodeActivity.class);
         intent.putExtras(bundle);
         startActivity(intent);                      //跳转到修改页面
